requirements.txt
pyadif
PyYAML
pymysql
oracledb
psycopg2
pyodbc
